server/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cors-headers==3.13.0
django-import-export==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environ==1.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.1.1
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tablib==3.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
